==> webscrapper/01_dexters_scrapper.ipynb <==
import sys
from importlib import reload
reload(sys)


import pandas as pd
import requests
from joblib import Parallel, delayed             
import pickle                                   
from tqdm import tqdm                            


def html_extractor(url):                         
    try:
        cookies = dict(cookies_are='working')
        r = requests.get(url, cookies=cookies)
        return r.text
    except:
        return "no html"

clickbaits_df = pd.read_csv('data/Clickbait.csv')

non_clickbaits_df = pd.read_csv('data/NonClickbait.csv')

clickbait_urls = clickbaits_df.status_link.values                    
non_clickbait_urls = non_clickbaits_df.status_link.values


clickbait_html = Parallel(n_jobs=40)(delayed(html_extractor)(u) for u in tqdm(clickbait_urls))    
pickle.dump(clickbait_html, open('data/clickbait_html.pkl', 'wb'), -1)

non_clickbait_html = Parallel(n_jobs=40)(delayed(html_extractor)(u) for u in tqdm(non_clickbait_urls))
pickle.dump(non_clickbait_html, open('data/non_clickbait_html.pkl', 'wb'), -1)
