==> webscrapper/03_dexters_scrapper.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split

clickbait_titles = pd.read_csv('data/Clickbait.csv')
non_clickbait_titles = pd.read_csv('data/NonClickbait.csv')

clickbait_features = pd.read_csv('data/clickbait_website_features.csv')
non_clickbait_features = pd.read_csv('data/non_clickbait_website_features.csv')

clickbait_entire = pd.concat([clickbait_titles, clickbait_features], axis=1)               
non_clickbait_entire = pd.concat([non_clickbait_titles, non_clickbait_features], axis=1)

clickbait_entire['label'] = 1
non_clickbait_entire['label'] = 0

fulldata = pd.concat([clickbait_entire, non_clickbait_entire],sort=True)                    
fulldata = fulldata.sample(frac=1).reset_index(drop=True)
fulldata = fulldata[fulldata.html_len != -1]

fulldata.to_csv('data/fulldata.csv', index=False)


df = pd.read_csv('data/fulldata.csv')


training_df, testing_df = train_test_split(df, random_state=42, test_size=0.1)

training_df.to_csv('data/train.csv', index=False, encoding='utf-8')
testing_df.to_csv('data/test.csv', index=False, encoding='utf-8')
