==> webscrapper/04_dexters_scrapper.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split

stop_words = ['site', 'navigation', 'new', 'times', 'york', 'information', 'index',
                      'like', 'related', 'search', 'follow', 'subscribe', 'subscribed', 'subscribing',
                       'spam', 'twitter', 'pinterest', 'facebook', 'google', 'privacy', 'policy', 'feedback',
                       'tweet', 'tweets', 'disclaimer', 'buzzfeed', 'clickhole', 'upworthy', 'cnn', 'nytimes',
                       'wikinews', 'instagram', 'newsletter', 'copyright', 'cnn.com', 'nytimes.com',
                       'buzzfeed.com', 'upworthy.com', 'clickhole.com']


def remove_stop_words(x):
    return ' '.join([word for word in str(x).lower().split() if word not in stop_words])

df = pd.read_csv('data/fulldata.csv')
df = df.drop_duplicates()
  
df.txt = df.textdata.apply(lambda x: str(x).replace('report an issue thanks', '').strip())

df.txt = df.txt.apply(remove_stop_words)
df.link = df.link_name.apply(remove_stop_words)



train_df, test_df = train_test_split(df, random_state=42, test_size=0.1)

train_df.to_csv('data/train.csv', index=False, encoding='utf-8')
test_df.to_csv('data/test.csv', index=False, encoding='utf-8')
