==> webscrapper/05_dexters_LSTM.ipynb <==
import pandas as pd
from keras.models import Sequential
from keras.layers import Dense, Activation, Flatten
from keras.layers.embeddings import Embedding
from keras.layers.recurrent import LSTM
from keras.layers.normalization import BatchNormalization
from keras.utils import np_utils
from keras.callbacks import ModelCheckpoint
from keras.layers.advanced_activations import PReLU
from keras.preprocessing import sequence, text
from keras.layers import SpatialDropout1D
from keras import metrics


train = pd.read_csv('data/train.csv')
test = pd.read_csv('data/test.csv')

y_train = train.label.values
y_test = test.label.values

tk = text.Tokenizer(num_words=200000)
train.link = train.link_name.astype(str)
test.link = test.link_name.astype(str)
train.text = train.textdata.astype(str)
test.text = test.textdata.astype(str)

max_len = 80

tk.fit_on_texts(list(train.link.values) + list(train.text.values) + list(test.link.values) + list(
    test.text.values))
x_train_title = tk.texts_to_sequences(train.link.values)
x_train_title = sequence.pad_sequences(x_train_title, maxlen=max_len)

x_train_textdata_01 = tk.texts_to_sequences(train.text.values)
x_train_textdata_01 = sequence.pad_sequences(x_train_textdata_01, maxlen=max_len)

x_test_title_01 = tk.texts_to_sequences(test.link.values)
x_test_title_01 = sequence.pad_sequences(x_test_title_01, maxlen=max_len)

x_test_textdata_02 = tk.texts_to_sequences(test.text.values)
x_test_textdata_02 = sequence.pad_sequences(x_test_textdata_02, maxlen=max_len)

word_index = tk.word_index
ytrain_enc = np_utils.to_categorical(y_train)

classifier = Sequential()
classifier.add(Embedding(len(word_index), 300, input_length=80, dropout=0.2),)
classifier.add(LSTM(300, return_sequences=True, dropout=0.2, recurrent_dropout=0.2))

classifier.add(Dense(200))
classifier.add(PReLU())
classifier.add(SpatialDropout1D(0.2))

classifier.add(BatchNormalization())

classifier.add(Dense(200))
classifier.add(PReLU())
classifier.add(SpatialDropout1D(0.2))
classifier.add(BatchNormalization())

classifier.add(Flatten())

classifier.add(Dense(2))
classifier.add(Activation('softmax'))


classifier.compile(loss='categorical_crossentropy', optimizer='adam', metrics=['acc' ])


checkpoint = ModelCheckpoint('data/weights.h5', monitor='val_acc', save_best_only=True, verbose=2)

classifier.fit(x_train_title, y=ytrain_enc,
                 batch_size=128, epochs=20, verbose=2, validation_split=0.1,
                 shuffle=True, callbacks=[checkpoint])
